GCS: Get rid of some compiler warnings.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@906 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/src/libs/opmapcontrol/src/core/rawtile.cpp
@@ -66,7 +66,7 @@ void RawTile::setZoom(const int &value)
 uint qHash(RawTile const& tile)
 {
     // RawTile tile=tilee;
-    quint64 tmp=(((quint64)(tile.zoom))<<54)+(((int)(tile.type))<<36)+(((quint64)(tile.pos.X()))<<18)+(((quint64)(tile.pos.Y())));
+    quint64 tmp=(((quint64)(tile.zoom))<<54)+(((quint64)(tile.type))<<36)+(((quint64)(tile.pos.X()))<<18)+(((quint64)(tile.pos.Y())));
   //  quint64 tmp5=tmp+tmp2+tmp3+tmp4;
     return ::qHash(tmp);
 }
